# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is test nodejs/express api for clone of auto-ria website",
    "version": "1.0.0",
    "title": " API",
    "contact": {
      "email": "anton.moskalenko.kh@gmail.com"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/ads/{userId}/all": {
      "get": {
        "tags": ["Advertisement"],
        "summary": "Get list active advertisements of specific user bases on his userId",
        "description": "This is open endpoint for getting active advertisements of specific user",
        "operationId": "getUserAds",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/AdsListResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "This user doesn't have any ads"
          }
        }
      }
    },
...
    "/ads/{adId}/images": {
      "post": {
        "tags": ["Advertisement"],
        "summary": "(testing only in postman) Upload images for advertisement",
        "description": "Closed endpoint for uploading images for advertisement. Testing only in postman, swagger does not want to accept image array",
        "operationId": "uploadImages",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [{"jwt_token": []}],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "adId",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the car ad to which the images will be uploaded."
            }
          },
          {
            "name": "files",
            "in": "formData",
            "description": "List of images to upload",
            "required": true,
            "type": "array",
            "items": {
              "type": "file"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Ad"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden. This user haven't access for this action"
          }
        }
      },
      "delete": {
        "tags": ["Advertisement"],
        "summary": "(testing only in postman) Delete images from advertisement",
        "description": "Closed endpoint for deleting images from advertisement. Testing only in postman, swagger does not want to accept image array",
        "operationId": "deleteImages",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [{"jwt_token": []}],
        "parameters": [
          {
            "in": "adId",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the car ad to which the images will be uploaded."
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Ad"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden. This user haven't access for this action"
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh pair of tokens",
        "description": "Closed endpoint for refresh tokens for user",
        "operationId": "refresh",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [{"jwt_token": []}],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Tokens"
            }
          },
          "401": {
            "description": "Unauthorised"
          }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Forgot password",
        "description": "Open endpoint for restore password of user",
        "operationId": "ForgotPassword",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "hello@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/auth/forgot-password/change": {
      "put": {
        "tags": ["Auth"],
        "summary": "Сhange forgot password",
        "description": "Closed endpoint with action token for change forgotten password of user",
        "operationId": "ForgotPassword",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "security": [{"jwt_token": []}],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "IamTestPassword&1111"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorised"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get list of active verified users",
        "description": "This is open endpoint for getting all active verified users",
        "operationId": "getAllUsers",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/UserListResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Users not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "jwt_token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Ad": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the car."
        },
        "brand": {
          "type": "string",
          "description": "Brand of the car."
        },
        "model": {
          "type": "string",
          "description": "Model of the car."
        },
        "year": {
          "type": "integer",
          "description": "Year of manufacture."
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "uri"
          },
          "description": "List of image URLs for the car."
        },
        "price": {
          "type": "number",
          "format": "float",
          "description": "Price of the car."
        },
        "description": {
          "type": "string",
          "description": "Description of the car."
        },
        "city": {
          "type": "string",
          "description": "City where the car is located."
        },
        "_userId": {
          "type": "string",
          "description": "Unique identifier for the user who create the advertisement."
        },
        "isActive": {
          "type": "boolean",
          "description": "Indicates if the car listing is active."
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp when the car listing was created."
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp when the car listing was last updated."
        },
        "message": {
          "type": "string",
          "description": "Optional message from moderator or admin"
        }
      },
      "required": [
        "_id",
        "brand",
        "model",
        "year",
        "price",
        "description",
        "city",
        "_userId",
        "isActive",
        "createdAt",
        "updatedAt"
      ]
    },
...
    "AdsListResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ad"
          },
          "description": "List of car objects."
        },
        "total": {
          "type": "integer",
          "description": "Total number of car listings."
        },
        "orderBy": {
          "type": "string",
          "description": "Field by which the results are ordered."
        },
        "order": {
          "type": "string",
          "description": "Order of the results, e.g., ascending or descending."
        },
        "limit": {
          "type": "integer",
          "description": "Number of items per page."
        },
        "page": {
          "type": "integer",
          "description": "Current page number."
        }
      },
      "required": [
        "data",
        "total",
        "orderBy",
        "order",
        "limit",
        "page"
      ]
    },
...
    "Tokens": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NmU5MTA1ODg3ZTQ0NzVkMzVjMGQ2MDEiLCJyb2xlIjoidXNlciIsImlhdCI6MTcyNjU1MDEwNSwiZXhwIjoxNzI2NTYwOTA1fQ.cU1WXGCg6FJrEaEqPFrC5FYkSAbR5oDXcjG95wOI1lM"
        },
        "refreshToken": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NmU5MTA1ODg3ZTQ0NzVkMzVjMGQ2MDEiLCJyb2xlIjoidXNlciIsImlhdCI6MTcyNjU1MDEwNSwiZXhwIjoxNzI2NjM2NTA1fQ.u4BDLOPi4RUZEXsL6qSznR-Ah64lm3wU2V8xZGX2HvQ"
        }
      },
      "required": [
        "accessToken",
        "refreshToken"
      ]
    },
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "66e2d8779a4fca98e3380fc4"
        },
        "name": {
          "type": "string",
          "example": "Anton23"
        },
        "email": {
          "type": "string",
          "example": "moderator@gmail.com"
        },
        "role": {
          "type": "string",
          "enum": [
            "user",
            "moderator",
            "admin"
          ],
          "example": "moderator"
        },
        "phone": {
          "type": "string",
          "example": "+380933453453"
        },
        "package": {
          "type": "string",
          "enum": [
            "base",
            "premium"
          ],
          "example": "premium"
        },
        "status": {
          "type": "string",
          "enum": [
            "active",
            "blocked"
          ],
          "example": "active"
        },
        "isVerified": {
          "type": "boolean",
          "example": true
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "example": "2024-09-12T12:03:03.740Z"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "example": "2024-09-12T12:03:03.740Z"
        }
      },
      "required": [
        "_id",
        "name",
        "email",
        "role",
        "phone",
        "package",
        "status",
        "isVerified",
        "createdAt",
        "updatedAt"
      ]
    },
...
    "UserListResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          },
          "description": "List of user objects."
        },
        "total": {
          "type": "integer",
          "description": "Total number of car listings."
        },
        "orderBy": {
          "type": "string",
          "description": "Field by which the results are ordered."
        },
        "order": {
          "type": "string",
          "description": "Order of the results, e.g., ascending or descending."
        },
        "limit": {
          "type": "integer",
          "description": "Number of items per page."
        },
        "page": {
          "type": "integer",
          "description": "Current page number."
        }
      },
      "required": [
        "data",
        "total",
        "orderBy",
        "order",
        "limit",
        "page"
      ]
    }
  }
}